Friend Request system is more polished
Allows user to send a request and immediately does not add the friend to the list of friends and when the other user accepts OR declines it adds it to the list of friends. But it does not add it to the list of friends of the person who sent it. --> need to fix